Fix login&logout button on Mypage Screen

+++ screens/Tabs/MyPageScreen.js
@@ -1,9 +1,17 @@
 import React, { useState, useEffect } from "react";
-import { StyleSheet, View, Text, AsyncStorage, Alert } from "react-native";
+import {
+  StyleSheet,
+  View,
+  Text,
+  AsyncStorage,
+  ScrollView,
+  Alert
+} from "react-native";
 import { Ionicons } from "@expo/vector-icons";
 import { Container } from "native-base";
 
 import Loader from "../../components/Loader";
+// import UserProfile from "../../components/UserProfile";
 import GhostButton from "../../components/Buttons/GhostButton";
 import { serverApi } from "../../components/API";
 import {
@@ -15,6 +23,8 @@ import {
 const MyPageScreen = ({ navigation }) => {
   const [loading, setLoading] = useState(false);
   const [profile, setProfile] = useState(null);
+  const [loggedIn, setLoggedIn] = useState(false);
+  const [buttonName, setButtonName] = useState(null);
   const preLoad = async () => {
     try {
       setLoading(true);
@@ -22,11 +32,14 @@ const MyPageScreen = ({ navigation }) => {
       if (userToken) {
         const userProfile = await serverApi.user(userToken);
         if (userProfile.data.data) {
-          // console.log(`userProfile: `, userProfile.data.data);
+          console.log(`userProfile: `, userProfile.data.data);
+          setLoggedIn(true);
+          setButtonName("임시 로그아웃 버튼");
           setProfile({ ...userProfile.data.data });
         }
       } else {
-        Alert.alert("로그인 해주세요");
+        setButtonName("임시 로그인 버튼");
+        // Alert.alert("로그인 해주세요");
       }
     } catch (e) {
       console.log(
@@ -38,30 +51,35 @@ const MyPageScreen = ({ navigation }) => {
   };
 
   const handleLogout = async () => {
-    const logout = await AsyncStorage.clear();
-    Alert.alert("로그아웃 되었습니다");
-    const resetAction = StackActions.reset({
-      index: 0,
-      actions: [NavigationActions.navigate({ routeName: "BottomNavigation" })]
-    });
-    navigation.dispatch(resetAction);
+    if (buttonName === "임시 로그아웃 버튼") {
+      const logout = await AsyncStorage.clear();
+      Alert.alert("로그아웃 되었습니다");
+      preLoad();
+      // const resetAction = StackActions.reset({
+      //   index: 0,
+      //   actions: [NavigationActions.navigate({ routeName: "Mypage" })]
+      // });
+      // navigation.dispatch(resetAction);
+    } else {
+      setButtonName("임시 로그아웃 버튼");
+      navigation.navigate("AuthNavigation");
+    }
   };
-
   useEffect(() => {
     preLoad();
   }, []);
+  // useEffect(() => {
+  //   preLoad();
+  // }, [profile]);
 
   return (
-    <Container>
+    <ScrollView>
       {loading ? (
         <Loader />
       ) : (
-        <View style={styles.Container}>
-          <Text>MyPageScreen</Text>
-          <GhostButton text={"임시 로그아웃 버튼"} onPress={handleLogout} />
-        </View>
+        <GhostButton text={buttonName} onPress={handleLogout} />
       )}
-    </Container>
+    </ScrollView>
   );
 };
